add controller manga
