+++ www/Controller/Main.class.php
@@ -17,4 +17,5 @@ class Main
     {
         $view = new View("contact");
     }
+    
 }
