+++ www/Controller/User.class.php
@@ -15,7 +15,8 @@ use App\Core\Session as Session;
 use App\Model\Manga;
 
 
-class User {
+class User 
+{
 
     public function login()
     {
@@ -270,79 +271,6 @@ class User {
         $view->assign("forum_data", $forum_data);
     }
 
-    public function manga()
-    {
-        $manga = new Manga();
-
-        if(!empty($_POST)) {
-
-            $result = Verificator::checkFormRegister($manga->getMangaForm(), $_POST);
-            print_r($result);
-
-            if (empty($result)) {
-                $manga->setTypeManga(htmlspecialchars($_POST["type"]));
-                $manga->setTitleManga(htmlspecialchars($_POST["title"]));
-                $manga->setDescriptionManga(htmlspecialchars($_POST["description"]));
-                $manga->setImageManga(htmlspecialchars($_POST["image"]));
-                $manga->setReleaseDateManga(htmlspecialchars($_POST["releaseDate"]));
-                $manga->setAuthorManga(htmlspecialchars($_POST["author"]));
-                $manga->setStatusManga(htmlspecialchars($_POST["status"]));
-                $manga->setCategoryManga(htmlspecialchars($_POST["category"]));
-                $manga->setNbTomesManga(htmlspecialchars($_POST["nbTomes"]));
-                $manga->setNbChaptersManga(htmlspecialchars($_POST["nbChapters"]));
-                $manga->setNbEpisodesManga(htmlspecialchars($_POST["nbEpisodes"]));
-                $manga->setDiffusionManga(htmlspecialchars($_POST["diffusion"]));
-                $manga->setNbSeasonsManga(htmlspecialchars($_POST["nbSeasons"]));
-                $manga->setProductionStudioManga(htmlspecialchars($_POST["productionStudio"]));
-                $manga->save();
-                echo "<script>alert('Votre manga a bien été mis à jour')</script>";
-            }
-        }
-        
-        $view = new View("manga", "back");
-        $view->assign("manga", $manga);
-
-        $manga_data = $manga->getMangas();
-        $view->assign("manga_data", $manga_data);
-    }
-
-    public function deleteManga()
-    {
-        $manga = new Manga();
-        if(!empty($_POST['manga_id'])){
-			$manga_Id = $_POST['manga_id'];
-            $manga->deleteManga($manga_Id);
-        }
-    }
-
-    public function editManga()
-    {
-        $manga = new Manga();
-        $view = new View("edit-manga", "back");
-        $view->assign("manga", $manga);
-        
-        if(!empty($_POST)){
-			$manga->setId($_GET["id"]);
-			$manga->setTypeManga(htmlspecialchars($_POST["type"]));
-            $manga->setTitleManga(htmlspecialchars($_POST["title"]));
-            $manga->setDescriptionManga(htmlspecialchars($_POST["description"]));
-            $manga->setImageManga(htmlspecialchars($_POST["image"]));
-            $manga->setReleaseDateManga(htmlspecialchars($_POST["releaseDate"]));
-            $manga->setAuthorManga(htmlspecialchars($_POST["author"]));
-            $manga->setStatusManga(htmlspecialchars($_POST["status"]));
-            $manga->setCategoryManga(htmlspecialchars($_POST["category"]));
-            $manga->setNbTomesManga(htmlspecialchars($_POST["nbTomes"]));
-            $manga->setNbChaptersManga(htmlspecialchars($_POST["nbChapters"]));
-            $manga->setNbEpisodesManga(htmlspecialchars($_POST["nbEpisodes"]));
-            $manga->setDiffusionManga(htmlspecialchars($_POST["diffusion"]));
-            $manga->setNbSeasonsManga(htmlspecialchars($_POST["nbSeasons"]));
-            $manga->setProductionStudioManga(htmlspecialchars($_POST["productionStudio"]));
-            $manga->save();
-            echo "<script>alert('Votre manga a bien été mis à jour')</script>";
-		}
-    }
-
-
     public function  recuperer_mdp()
     {
         $user = new UserModel();
@@ -356,7 +284,7 @@ class User {
         $view->assign("errors", $errors);
     }
 
-    public function  initialiser_mdp()
+    public function initialiser_mdp()
     {
         $user = new UserModel();
         $mdp = new PasswordModel();
