+++ www/routes.yml
@@ -99,17 +99,17 @@
   security: [Admin,Editeur,Abonne]
 
 /manga:
-  controller: user
+  controller: manga
   action: manga
   security: [Admin,Editeur,Abonne]
 
 /editManga:
-  controller: user
+  controller: manga
   action: editManga
   security: [Admin,Editeur,Abonne]
 
 /deleteManga:
-  controller: user
+  controller: manga
   action: deleteManga
   security: [Admin]
